Sort interface lists in stubs
While the order of interfaces in the list does not matter in the stub
file this change enforces a consistent order independent of the source.
That ensures that code changes which change the order do not cause
unnecessary churn in the generated stubs.

This uses the same order as is used in signature files that do not use
the legacy order.

Bug: 348701068
Test: ./gradlew
Change-Id: Ie246845a09c4bd9ef650107062b5f58acddee40a