Remove TypeItem.toTypeString(...)'s filter parameter
Previously, the `filter` parameter was used by `SignatureWriter` and
`JavaStubWriter` but they no longer use it as that work is being done
by `FilteringApiVisitor`. So, this change removes it.

Rather than remove the tests in `CommonTypeStringTest` that verified
this behavior and lose the test coverage this change reimplements the
tests. the `TypeTransformer` used by `FilteringApiVisitor` to do the
equivalent work was extracted and generalized and is now used by the
test to basically replicate the previous behavior. That improves the
test coverage of the `TypeTransformer`.

Bug: 348701068
Test: ./gradlew
Change-Id: I851101abdedf43c79ba6069778fadfb7f8150548