Make JDiffXmlWriter extend BaseItemVistor
Previously, `JDiffXMLWriter` extended `ApiVisitor`. This change
switches it to extend `BaseItemVisitor` instead. The existing behavior
is maintained by creating a `FilteringApiVisitor` that delegates to the
`JDiffXMLWriter`.

At the moment the `JDiffXMLWriter` still does its own filtering but a
follow-up change will it to take advantage of the filtering done by
`FilteringApiVisitor`.

It was necessary to delegate calls to `visitCodebase(Codebase)` and
`afterVisitCodebase(Codebase)` to `FilteringApiVisitor.delegate` as
`JDiffXMLWriter` uses that to generate the open and close tags for the
root `api` element.

Bug: 348701068
Test: ./gradlew
Change-Id: Iceab6a7e89377b41bd0789bf35a00afa0eb4d6b9