Optimize interface types list handling in FilteringApiVisitor
Previously, the interface types list would always be sorted and then
have their type use annotations filtered irrespective of whether it was
actually needed.

This change optimizes the behavior based on the size of the filtered
list:
1. If the list is empty then sorting and type use annotation filtering
   are not needed so it returns early.
2. If the list only has 1 element then sorting is unnecessary but
   type use annotation filtering is still needed.
3. If the list has 2 or more elements then sorting and type use
   annotation filtering are all needed.

As a result `interfaceListSorter` is only called with a filtered list
containing at least 2 items which allows the empty check in
`SignatureWriter` to be removed.

Bug: 348701068
Test: ./gradlew
Change-Id: Icf49f4a96116e7e6ec94460f683da8c9a13b1821