Add special DelegatedVisitor for use with FilteringApiVisitor
Previously, both the `FilteringApiVisitor` and its `BaseItemVisitor`
delegates were supplied with the `BaseItemVisitor` parameters, i.e.
`visitConstructorsAsMethods` and `nestInnerClasses`. However, they do
not need both parameters.

`FilteringApiVisitor` needs `nestInnerClasses` as that controls the
interleaving of `visitClass(...)` and `afterVisitClass(...)` calls that
`FilteringApiVisitor` makes to the delegate. However, it does not need
`visitConstructorsAsMethods` as that is implemented in the
`visitConstructor(...)` and `afterVisitConstructor(...)` methods which
`FilteringApiVisitor` overrides.

The delegates do not need  `visitConstructorsAsMethods`, they all pass
`false` for that. They also do not need `nestInnerClasses` as that is
handled by `FilteringApiVisitor`.

This change removes the `visitConstructorsAsMethods` from the
`FilteringApiVisitor` constructor and adds a special `DelegatedVisitor`
interface that just provides the `visit...()` methods needed by the
delegates. It migrates the existing delegates to use the new interface.

As `StubWriter` itself delegates to `JavaStubWriter` they both were
changed to implement `DelegatedVisitor` instead of `BaseItemVisitor`.

Bug: 348701068
Test: ./gradlew
Change-Id: Ic7448011821aaeafb8a52bd9b15ceb0025259673