Remove ApiVisitor.includeEmptyOuterClasses
Previously, the `includeEmptyOuterClasses` was used in the following
expression `includeEmptyOuterClasses || shouldEmitClassBody(vc)`.
So, it only had an effect if it was `true`.

Removing it is only safe if `shouldEmitClassBody(vc)` will always
return `true` in any situation when `includeEmptyOuterClasses` was
`true`.

It was only set to `true` in two places:
1. `StubWriter` which generates stubs.
2. `MarkPackagesAsRecent` which replaces `Nullable`/`NonNull` with
   `RecentlyNullable`/`RecentlyNonNull` respectively; the latter are
   only written out into stubs.

The only way an empty outer class can occur is if there is say a class
`Foo` that is part the public API (unannotated) which which has a
nested class `Foo.Bar` that is part of the system API (annotated).
When generating the system API the `Foo` class will be empty. That
behavior is used when generating signature files to avoid generating an
empty `Foo` class in the `system-current.txt` file.

However, when generating stubs for system it has to include all the
classes from public, including `Foo`. So, for the purposes of stub
generation `Foo` is considered part of the system API. That means that
`filterEmit` will return `true` when called on it as `filterEmit` for
stubs sets `includeApisForStubPurposes=true`. That in turn means that
`shouldEmitClassBody(vc)` will return `true` because it looks like
this:
```
    when {
        filterEmit.test(vc.cls) -> true
        vc.nonEmpty() -> filterReference.test(vc.cls)
        else -> false
    }
```

Bug: 350130387
Test: ./gradlew
      # Extensive testing was done on Android to ensure that it does
      # not affect the generated stubs.
Change-Id: Ief81179eae932a9655c3424984f3718bf5ee86cc