Remove workaround for writing partial signature file to JDiff
Change https://r.android.com/874370 added a workaround for when partial
signature files that looked something like this were converted to
JDiff:
```
package test.pkg {
  public class Foo.Bar {...}
}
```

When it was loaded `Foo` would have `emit=false` (because it was not
defined in the file). That meant that when recursing down through the
classes to generate the JDiff file it would stop at `Foo` and not
visit `Bar`. Change https://r.android.com/874370 fixed that by
deferring the `emit` check until after the `Foo` had been visited if
`preFiltered = true`.

However, that is no longer needed as the `JDiffWriter`'s `ApiVisitor`
sets `nestInnerClass = false` which means they are flattened and are
visited as if they are top level classes in the package. That means
it no longer visits `Foo` in order to get to `Foo.Bar`.

Test `JDiffXmlTest.Test partial signature files` was added in
https://r.android.com/874370 and that still passes.

Bug: 350130387
Test: ./gradlew
Change-Id: Ib408c1a40042fe01d6deef560ab20095c67f096a