Inline include(VisitCandidate) dedup include(ClassItem) call
The `ApiVisitor.include(VisitCandidate)` is only called from one place,
i.e. `VisitCandidate.accept()` so this change inlines it resulting in
something like:
```
    fun accept() {
        if (!include(cls) || !shouldEmitClassBody(this)) {
            return
        }
	...
```

However `accept()` is only called from `visit(ClassItem)` which looks
something like this:
```
    override fun visit(cls: ClassItem) {
        if (!include(cls)) {
            return
        }

        // ...
        val candidate = VisitCandidate(cls)
        candidate.accept()
    }
```

So, the `include(cls)` call in `VisitCandidate.accept()` is always
going to return `true` so this change removes it.

Bug: 350130387
Test: ./gradlew
Change-Id: Ib83a137a04e9356f78078c5879642148d49e5d87