Stop initializing VisitCandidate's emittable members properties lazily
Previously, `VisitCandidate` would initialize its emittable members
properties lazily because that is expensive and they may not have been
needed because it may have been discarded before those properties were
accessed. However, the preceding change means that a `VisitCandidate`
is only created when it is needed so there is no need to pay the cost
of doing it lazily.

This also adds some private extension functions to remove some of the
duplication of the properties.

Bug: 350130387
Test: ./gradlew
Change-Id: Idb9c528d46383c043779f1aeff87a76fae5bd639