Move TextModelSuiteRunner into testFixtures
Testing the snapshot code requires being able to run an existing
`ModelSuiteRunner` in a new Gradle project. The simplest model to start
with is the text model so this makes its `ModelSuiteRunner` available
for use in another project's tests by adding it to the `testFixtures`
which can be shared with another project's tests.

Bug: 353737744
Test: ./gradlew
Change-Id: I952b61b5968b696802936300ca21de35da2d8f3f