Add simple CodebaseSnapshotTaker
A simple class that creates a new `Codebase` that is a snapshot of the
supplied one, but does not copy any of the contents.

This sets up the visitor infrastructure so that the `Codebase` which is
being copied will be visited and then delegate to
`CodebaseSnapshotTaker` through a `DelegatedVisitor` interface. That
will allow it to be used with a `FilteringApiVisitor` in future.
However, for now it adds a new `NonFilteringDelegatingVisitor` that
will just visit the whole input `Codebase` without filtering.

Bug: 353737744
Test: ./gradlew
Change-Id: I775ea144274cd3b2cdba9f5bba8293b347490c5d