Restructure Turbine package creation to avoid mutating originallyHidden
Previously, Turbine would interleave the creation of packages
associated with a `package-info.java` file and packages associated with
a class. That meant that a package might be created when it saw a class
in that package (when information from `package-info.java`, e.g.
documentation, was not available) and then have to update it later. As
the `originallyHidden` property depended on `documentation` it needed
to be updated.

This change restructures that so that all the `package-info.java` files
are processed first, then the classes and then, if necessary a root
package is added. That avoids the need to update the `originallyHidden`
property so the `updateOriginallyHiddenStatus(String)` method was no
longer used and was removed.

Bug: 350927039
Test: ./gradlew
Change-Id: I5f374d324f89cb32c7f23d4d37e339b194e1b558