Make TurbineItem.originallyHidden consistent with PsiItem's
Switches the order of `hidden` and `originallyHidden` to match
`PsiItem`'s and adds a check for `@suppress`. The latter was added to
support Kotlin but there is no restriction on it being used in Java as
well so this change removes one possible, albeit minor, source of
differences between Turbine and Psi.

Bug: 351118278
Test: ./gradlew
Change-Id: I12a44f68d4d7068c6b927e47b4113be1aa2d161f