Add ItemLanguage to allow DefaultItem to implement isJava/isKotlin
`TextItem` implements `isJava()` and `isKotlin()` to throw an error as
it does not know whether a specific element is java or kotlin. However,
`TurbineItem` always returns `true` for `isJava()` as Turbine only
supports Java. To allow them to be combined this change adds an
`ItemLanguage` property to `DefaultItem` that supports Java, Kotlin and
unknown languages. The latter is used by `TextItem` to replicate its
previous behavior.

Bug: 351410134
Test: ./gradlew
Change-Id: I4465e06a1b6c64f86e986374723205e11f4e35b8